add new routes for EPI categories, equipment, suppliers, brands, and models; remove old equipment routes
