+++ app/routes/EPI/__init__.py
@@ -1,0 +1,6 @@
+from .categorias import categoria
+from .equipamentos import equip
+from .fornecedores import fornecedor
+from .modelos import modelo
+
+__all__ = ["modelo", "categoria", "fornecedor", "equip", "categoria"]
