+++ app/routes/EPI_old/equipamentos.py
