+++ app/routes/EPI_old/estoque.py
